### 2024-05-09 #6738 @editorialbot
[PRE REVIEW]: squishyplanet: modeling transits of non-spherical exoplanets in JAX
**Submitting author:** @ben-cassese (<a href="http://orcid.org/0000-0002-9544-0118">Benjamin Cassese</a>)
**Repository:** https://github.com/ben-cassese/squishyplanet
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
