+++ README.md
@@ -1,6 +1,6 @@
 # squishyplanet
 
-
+[![status](https://joss.theoj.org/papers/2b62acea4e4d6c0de73309184861d1e9/status.svg)](https://joss.theoj.org/papers/2b62acea4e4d6c0de73309184861d1e9)
 [![Documentation Status](https://readthedocs.org/projects/squishyplanet/badge/?version=latest)](https://squishyplanet.readthedocs.io/en/latest/?badge=latest)
 ![Build Status](https://github.com/ben-cassese/squishyplanet/actions/workflows/tests.yml/badge.svg)
 [![Coverage Status](https://coveralls.io/repos/github/ben-cassese/squishyplanet/badge.svg?branch=main&service=github)](https://coveralls.io/github/ben-cassese/squishyplanet?branch=main)
