### 2024-06-14 72d29e6
initial visualization function
