### 2024-06-21 244599f
s vector implementation
