+++ squishyplanet/engine/spherical_harmoninc_basis_transform.py
@@ -1,0 +1,225 @@
+# Taken directly from the repository associated with
+# Luger et al. 2019 (doi:10.3847/1538-3881/aae8e5)
+# Many thanks to the authors for making this code available!
+
+import jax
+
+jax.config.update("jax_enable_x64", True)
+import jax.numpy as jnp
+
+import numpy as np
+import sympy
+from sympy import symbols, floor, sqrt, pi, factorial, zeros
+from sympy.functions.special.tensor_functions import KroneckerDelta
+
+
+# Define our symbols
+x, y, z, mu, nu, l, m, j, k, p, q, n = symbols("x y z mu nu l m j k p q n")
+
+
+# Definitions from `bp.ipynb`
+def ptilde(n, x, y):
+    """Return the n^th term in the polynomial basis."""
+    l = floor(sqrt(n))
+    m = n - l * l - l
+    mu = l - m
+    nu = l + m
+    if nu % 2 == 0:
+        i = mu // 2
+        j = nu // 2
+        k = 0
+    else:
+        i = (mu - 1) // 2
+        j = (nu - 1) // 2
+        k = 1
+    return x**i * y**j * sqrt(1 - x**2 - y**2) ** k
+
+
+# Definitions from `ylmxyz.ipynb`
+def A(l, m):
+    """A spherical harmonic normalization constant."""
+    return sqrt(
+        (2 - KroneckerDelta(m, 0))
+        * (2 * l + 1)
+        * factorial(l - m)
+        / (4 * pi * factorial(l + m))
+    )
+
+
+def B(l, m, j, k):
+    """Another spherical harmonic normalization constant."""
+    try:
+        ratio = factorial((l + m + k - 1) / 2) / factorial((-l + m + k - 1) / 2)
+    except ValueError:
+        ratio = 0
+    return (
+        2**l
+        * factorial(m)
+        / (factorial(j) * factorial(k) * factorial(m - j) * factorial(l - m - k))
+        * ratio
+    )
+
+
+def C(p, q, k):
+    """Return the binomial theorem coefficient `C`."""
+    return factorial(k / 2) / (
+        factorial(q / 2) * factorial((k - p) / 2) * factorial((p - q) / 2)
+    )
+
+
+def Y(l, m, x, y):
+    """Return the spherical harmonic of degree `l` and order `m`."""
+    res = 0
+    z = sqrt(1 - x**2 - y**2)
+    if m >= 0:
+        for j in range(0, m + 1, 2):
+            for k in range(0, l - m + 1, 2):
+                for p in range(0, k + 1, 2):
+                    for q in range(0, p + 1, 2):
+                        res += (
+                            (-1) ** ((j + p) // 2)
+                            * A(l, m)
+                            * B(l, m, j, k)
+                            * C(p, q, k)
+                            * x ** (m - j + p - q)
+                            * y ** (j + q)
+                        )
+            for k in range(1, l - m + 1, 2):
+                for p in range(0, k, 2):
+                    for q in range(0, p + 1, 2):
+                        res += (
+                            (-1) ** ((j + p) // 2)
+                            * A(l, m)
+                            * B(l, m, j, k)
+                            * C(p, q, k - 1)
+                            * x ** (m - j + p - q)
+                            * y ** (j + q)
+                            * z
+                        )
+    else:
+        for j in range(1, abs(m) + 1, 2):
+            for k in range(0, l - abs(m) + 1, 2):
+                for p in range(0, k + 1, 2):
+                    for q in range(0, p + 1, 2):
[…]
