+++ squishyplanet/engine/spherical_harmonics.py
@@ -110,19 +110,19 @@ def _big_G(n, x, y, z):
         ((nu % 2 == 1) & (mu == 1) & (l % 2 == 0)), x ** (l - 2) * z**3, big_G_x
     )
     big_G_x = jnp.where(
-        ((nu % 2 == 1) & (mu == 1) & (l % 2 == 1) & (big_G_x == 0)),
+        ((nu % 2 == 1) & (mu == 1) & (l % 2 == 1) & ~((l == 1) & (m == 0))),
         x ** (l - 3) * y * z**3,
         big_G_x,
     )
 
     # only 3/5 cases y is non-zero
     big_G_y = 0
-    big_G_y = jnp.where(nu % 2 == 0, x ** ((mu + 2) / 2) * y ** (nu / 2), big_G_y)
+    big_G_y = jnp.where((nu % 2 == 0), x ** ((mu + 2) / 2) * y ** (nu / 2), big_G_y)
     big_G_y = jnp.where(
-        (l == 1) & (m == 0), (1 - z**3) / (3 * (1 - z**2)) * (x), big_G_y
+        ((l == 1) & (m == 0)), (1 - z**3) / (3 * (1 - z**2)) * (x), big_G_y
     )
     big_G_y = jnp.where(
-        (big_G_x == 0) & (big_G_y == 0),
+        (~(nu % 2 == 0)) & (~((l == 1) & (m == 0))),
         x ** ((mu - 3) / 2) * y ** ((nu - 1) / 2) * z**3,
         big_G_y,
     )
@@ -149,3 +149,25 @@ def big_G(n, x, y, z):
 
     """
     return jax.vmap(_big_G, in_axes=(0, None, None, None))(n, x, y, z)
+
+
+# for debugging: this should matche map.flux() in starry
+# def wrapper(alpha):
+#     a = 1
+#     b = 1
+
+#     x = a*jnp.cos(alpha)
+#     y = b*jnp.sin(alpha)
+#     # z = jnp.sqrt(1 - x**2 - y**2)
+#     z = 0
+
+#     dx = -jnp.sin(alpha)*a
+#     dy = jnp.cos(alpha)*b
+
+#     return jnp.sum(big_G(jnp.arange(9), x, y, z) * jnp.array([dx, dy]), axis=1)
+
+# s_vec = quadgk(wrapper, jnp.array([0, 2*jnp.pi]), epsabs=1e-12, epsrel=1e-12)[0]
+# af = generate_spherical_harmonic_basis_transform_matrix(2)
+# v = jnp.zeros(9)
+# v = v.at[2].set(1)
+# s_vec.T @ af @ v * 2 / jnp.sqrt(jnp.pi)
