### 2024-07-29 ben-cassese/squishyplanet#10 @rferrerc
Initial comments on JOSS submission
Hi @ben-cassese et al!

I'm one of the reviewers for the JOSS submission; I have a few initial comments (mostly on the "Documentation" and "Software paper" sections of the checklist) that I figured might be easier to share now, while I finish taking a look at the rest of the submission.

**Functionality:**

- Installation: I use a Linux machine, and following the [installation instructions](https://squishyplanet.readthedocs.io/en/latest/installation.html#mac-linux-users) I tried all three options; options 1 and 3 worked well out-of-the-box, but trying option 2 I got the following error: `ERROR: Invalid requirement: 'squishyplanet@git:https://github.com/ben-cassese/squishyplanet' Hint: It looks like a path. File 'squishyplanet@git:https://github.com/ben-cassese/squishyplanet' does not exist.` . Not sure if the command line showed on option 2 is also intended to be an out-of-the-box line to be run or if previous steps are implied for the source installation, but flagging this as some clarification would be useful.

**Documentation:**

- Example usage: Most of the core methods are amply used in the documentation examples, except for OblateSystem.loglike(). I believe it would be useful to show an example using this method on real or simulated data, since it would also help satisfy the checklist requirement of "including examples of how to use the software (_ideally to solve real-world analysis problems_)".
- Functionality documentation: I think it would be helpful to define the conventions used for some of the orbital elements (e.g., is an "edge-on" orbit i=0° or i=90°, in which direction does big Omega increase, etc). Either in the API docs or in the geometry visualization section would be helpful.

**Software paper:**

- State of the field: In the second paragraph of the Statement of Need section, it would be helpful to have a few words on how `squishyplanet` differs from `ellc` (the code from Maxted 2016) since, from what I understand in the text, Maxted 2016 can also model non-spherical transits.

I will continue to look at the code and text in the upcoming days, and feel free to reach out for any clarification on the comments above.

Thanks!

[Review issue](https://github.com/openjournals/joss-reviews/issues/6972#issue-2401368914)
