+++ squishyplanet/engine/development/quadratic_limb_darkened_transit.py
@@ -1,265 +1,0 @@
-# this is a quadratic-only limb darkening model, uses a different method than the
-# Agol, Luger, Foreman-Mackey 2020 scheme. Should produce the same results as the
-# more generate polynomial_limb_darkened_transit.py model when all terms > u2 are 0
-
-import jax
-
-jax.config.update("jax_enable_x64", True)
-import jax.numpy as jnp
-
-from quadax import quadgk
-
-from squishyplanet.engine.planet_3d import planet_3d_coeffs
-from squishyplanet.engine.planet_2d import planet_2d_coeffs
-from squishyplanet.engine.parametric_ellipse import (
-    poly_to_parametric,
-    cartesian_intersection_to_parametric_angle,
-)
-from squishyplanet.engine.skypos import skypos
-from squishyplanet.engine.star_planet_intersection import single_intersection_points
-from squishyplanet.engine.kepler import kepler
-
-
-@jax.jit
-def planet_flux_integrand(alpha, u1, u2, c_x1, c_x2, c_x3, c_y1, c_y2, c_y3, **kwargs):
-    return (
-        (
-            c_x3 * (-(jnp.sin(alpha) * c_y1) + jnp.cos(alpha) * c_y2)
-            - c_x2 * (c_y1 + jnp.cos(alpha) * c_y3)
-            + c_x1 * (c_y2 + jnp.sin(alpha) * c_y3)
-        )
-        * (
-            3
-            * u2
-            * (
-                (jnp.cos(alpha) * c_x1 + jnp.sin(alpha) * c_x2 + c_x3) ** 2
-                + (jnp.cos(alpha) * c_y1 + jnp.sin(alpha) * c_y2 + c_y3) ** 2
-            )
-            ** 2
-            - 4
-            * (u1 + 2 * u2)
-            * (
-                -1
-                + jnp.sqrt(
-                    1
-                    - (jnp.cos(alpha) * c_x1 + jnp.sin(alpha) * c_x2 + c_x3) ** 2
-                    - (jnp.cos(alpha) * c_y1 + jnp.sin(alpha) * c_y2 + c_y3) ** 2
-                )
-            )
-            + 2
-            * (
-                (jnp.cos(alpha) * c_x1 + jnp.sin(alpha) * c_x2 + c_x3) ** 2
-                + (jnp.cos(alpha) * c_y1 + jnp.sin(alpha) * c_y2 + c_y3) ** 2
-            )
-            * (
-                3
-                - 3 * u1
-                - 6 * u2
-                + 2
-                * u1
-                * jnp.sqrt(
-                    1
-                    - (jnp.cos(alpha) * c_x1 + jnp.sin(alpha) * c_x2 + c_x3) ** 2
-                    - (jnp.cos(alpha) * c_y1 + jnp.sin(alpha) * c_y2 + c_y3) ** 2
-                )
-                + 4
-                * u2
-                * jnp.sqrt(
-                    1
-                    - (jnp.cos(alpha) * c_x1 + jnp.sin(alpha) * c_x2 + c_x3) ** 2
-                    - (jnp.cos(alpha) * c_y1 + jnp.sin(alpha) * c_y2 + c_y3) ** 2
-                )
-            )
-        )
-    ) / (
-        12.0
-        * (
-            (jnp.cos(alpha) * c_x1 + jnp.sin(alpha) * c_x2 + c_x3) ** 2
-            + (jnp.cos(alpha) * c_y1 + jnp.sin(alpha) * c_y2 + c_y3) ** 2
-        )
-    )
-
-
-# @jax.jit
-def star_flux_integral(alpha1, alpha2, u1, u2, c_x1, c_x2, c_x3, c_y1, c_y2, c_y3):
[…]
