### 2024-07-30 bb6d386
fix install instructions
