+++ docs/installation.md
@@ -15,7 +15,7 @@ python -m pip install squishyplanet
 <span style="font-size:larger;">Option 2: install from source:</span>
 
 ```bash
-python -m pip install squishyplanet@git:https://github.com/ben-cassese/squishyplanet
+python -m pip install git+https://github.com/ben-cassese/squishyplanet
 ```
 
 <span style="font-size:larger;">Option 3: clone and install an editable version:</span>
