### 2024-08-02 ben-cassese/squishyplanet#10 @catrionamurray
Hi! I'm the second reviewer for this JOSS submission. I'm currently creating a separate GitHub issue with my comments but I just wanted to +1 on all @rferrerc's comments and add that I also have an issue with only the second installation method on my MacBook M1:
```
python -m pip install squishyplanet@git:https://github.com/ben-cassese/squishyplanet
Collecting squishyplanet@ git:///https://github.com/ben-cassese/squishyplanet
ERROR: Could not install packages due to an OSError: No connection adapters were found for 'git:///https://github.com/ben-cassese/squishyplanet'
```
