### 2024-08-02 ben-cassese/squishyplanet#11 @catrionamurray
JOSS Review 
Hi @ben-cassese and everyone!

I'm the other reviewer for JOSS review: https://github.com/openjournals/joss-reviews/issues/6972. I'm in the process of testing the functionality but here are some initial comments on the docs, paper and installation:

**General Checks**
- Reproducibility: Not quite original results, however, the figure in the paper is only briefly described with "a more complete description" in the online documentation. I followed the link to the documentation, however, this specific plot is not generated in the docs. I would like to see the plot + full explanation of each light curve in the docs and/or in the paper.

**Functionality**
- Installation: Similar to https://github.com/ben-cassese/squishyplanet/issues/10 I also have an issue with the second installation method. The 1st and 3rd installation methods work fine for me.
Error:
```
python -m pip install squishyplanet@git:https://github.com/ben-cassese/squishyplanet
Collecting squishyplanet@ git:///https://github.com/ben-cassese/squishyplanet
ERROR: Could not install packages due to an OSError: No connection adapters were found for 'git:///https://github.com/ben-cassese/squishyplanet'
```

**Documentation**
- Example usage: The Quickstart tutorial has no documentation. Since this is the user's first introduction to the code it would be nice to have some context and information about running the code here!
- Functionality documentation: I think the API has a good depth of detail, however, in general, descriptions of individual steps in the tutorials lack clarity. Some of this would be remedied by a more detailed Quickstart tutorial. Some examples to better illustrate my point:
_Illustrations_ -> "_Now let’s flatten it:_" - How do you do this? What axis are you flattening? What parameter(s) are important here?
_Illustrations_ -> "_That was all for emitted light, but we can visualize the reflected light too:_" - What parameter(s) are you changing to do this?
_Create a Phase Curve_ -> No description of the setup under Reflections.
- The `loglike` function is not demonstrated in the tutorials.

**Software Paper**
- State of the field: added small follow-up Q to https://github.com/ben-cassese/squishyplanet/issues/10 on oblateness of host stars.
