### 2024-08-05 ben-cassese/squishyplanet#10 @ben-cassese
Hi both!

Thanks for the initial comments! I've started working on these fixes and will let you know when they're ready for another look. In the meantime, just wanted to acknowledge that I'm digging into them this week 🙂
