### 2024-08-06 9fa8014
switched projected r parameterization in 2d case
