+++ tests/test_oblate_system.py
@@ -14,7 +14,7 @@ def test_limb_darkening_profile_helper():
         key, *subkeys = jax.random.split(key, (3,))
         poly_order = jax.random.randint(subkeys[0], (1,), minval=2, maxval=7)
         poly_order = int(poly_order[0])
-        u_coeffs = jax.random.uniform(key, (poly_order,), minval=0.0, maxval=1.0)
+        u_coeffs = jax.random.uniform(subkeys[1], (poly_order,), minval=0.0, maxval=1.0)
 
         state = {
             "t_peri": 0.0,
@@ -165,7 +165,7 @@ def test_lightcurve():
         "compute_stellar_ellipsoidal_variations": False,
         "compute_stellar_doppler_variations": False,
         "parameterize_with_projected_ellipse": True,
-        "projected_r": 0.1,
+        "projected_effective_r": 0.1,
         "projected_f": 0.1,
         "projected_theta": 0.1,
         "phase_curve_nsamples": 10,
@@ -305,7 +305,7 @@ def test_loglike():
         "compute_stellar_ellipsoidal_variations": True,
         "compute_stellar_doppler_variations": True,
         "parameterize_with_projected_ellipse": False,
-        "projected_r": 0.0,
+        "projected_effective_r": 0.0,
         "projected_f": 0.0,
         "projected_theta": 0.0,
         "phase_curve_nsamples": 100,
