### 2024-08-06 b9c389c
paper edits, fit tutorial draft, added illustrations
