+++ paper.md
@@ -48,7 +48,7 @@ While astronomers often assume that exoplanets are perfect spheres when analyzin
 
 The study of exoplanets, or planets that orbit stars beyond the sun, is a major focus of the astronomy community. Many of these studies center on the analysis of time series photometric (or spectroscopic) observations collected when a planet happens to pass through the line of sight between an observer and its host star. By modeling the fraction of starlight intercepted by the coincident planet, astronomers can deduce basic properties of the system such as the planet's relative size, its orbital period, and its orbital inclination.
 
-The past 20 years have seen extensive work both on theoretical model development and computationally efficient implementations of these models. Notable examples include @mandel_agol, @batman, and @exoplanet, though many other examples can be found. Though each of these packages make different choices, the majority of them (with notable exceptions, including @ellc) do share one common assumption: the planet under examination is a perfect sphere.
+The past 20 years have seen extensive work both on theoretical model development and computationally efficient implementations of these models. Notable examples include @mandel_agol, @batman, and @exoplanet, though many other examples can be found. Though each of these packages make different choices, the majority of them (with notable exceptions, including @ellc[^1]) do share one common assumption: the planet under examination is a perfect sphere.
 
 This is both a reasonable and immensely practical assumption. It is reasonable because firstly, a substantial fraction of planets, especially rocky planets, are likely quite close to perfect spheres (Earth's equatorial radius is only 43 km greater than its polar radius, a difference of 0.3%). Secondly, at the precision of most survey datasets (e.g. *Kepler* and *TESS*), even substantially flattened planets would be nearly indistinguishable from a spherical planet with the same on-sky projected area [@zhu2014]. It is practical since, somewhat miraculously, this assumption enables an analytic solution for the amount of flux blocked by the planet at each timestep. This is true even if the intensity of the stellar surface varies radially according to a nearly arbitrarily complex polynomial [@alfm].
 
@@ -63,6 +63,9 @@ The amplitudes of these effects are quite small compared to the full depth of th
 We leave a detailed description of the mathematics and a corresponding series of visualizations for the online documentation. There we also include confirmation that our implementation, when modeling the limiting case of a spherical planet, agrees with previous well-tested models even for high-order polynomial limb darkening laws. More specifically, we show that that lightcurves of spherical planets generated with `squishyplanet` deviate by no more than 100 ppb from those generated with  `jaxoplanet` [@jaxoplanet], the `JAX`-based rewrite of the popular transit modeling package `exoplanet` [@exoplanet] that also implements the arbitrary-order polynomial limb darkening algorithm presented in @alfm. Finally, we demonstrate `squishyplanet`'s limited support for phase curve modeling. 
 
 We hope that a publicly-available, well-documented, and highly accurate model for non-spherical transiting exoplanets will enable thorough studies of planets' shapes and lead to more data-informed constraints on their interior structures.
+
[…]
