### 2024-08-11 728d9d6
ld coeffs from stellar grid profile
