+++ squishyplanet/__init__.py
@@ -4,4 +4,4 @@ __all__ = ["OblateSystem"]
 __url__ = "https://github.com/ben-cassese/squishyplanet"
 __license__ = "MIT"
 __description__ = "Modeling transits of non-spherical exoplanets"
-__version__ = "0.1.0"
+__version__ = "0.1.2"
