### 2024-08-12 d9f8ce3
log_jitter var->std, misc tutorial edits
