+++ paper.md
@@ -64,7 +64,7 @@ We leave a detailed description of the mathematics and a corresponding series of
 
 We hope that a publicly-available, well-documented, and highly accurate model for non-spherical transiting exoplanets will enable thorough studies of planets' shapes and lead to more data-informed constraints on their interior structures.
 
-[^1]: Though `ellc`, and `squishyplanet` share the same goal of modeling transits of non-spherical planets, they differ in a few key ways. First, `ellc` requires users to select from a set of predefined limb darkening laws, while `squishyplanet` allows for any law that can be cast as a polynomial (e.g. high-order approximations to grid-based models). Second, `ellc` allows for gravity-deformed stars, while `squishyplanet` always models the central star as a sphere and restricts triaxial deformations to the planet only. Third, `ellc` allows users to model radial velocity curves, including the Rossiter-McLaughlin effect, while `squishyplanet` is focused on lightcurve modeling only. In terms of implementation, `ellc` is written in Fortran and wrapped in Python, while `squishyplanet` is written in `JAX`. Also, `ellc` integrates the flux blocked by the planet via 2D numerical integration, while `squishyplanet` uses a 1D numerical integration scheme. We believe that these tools will be complementary and that users will benefit from having both available.
+[^1]: Though `ellc`, and `squishyplanet` share the same goal of modeling transits of non-spherical planets, they differ in a few key ways. First, `ellc` requires users to select from a set of predefined limb darkening laws, while `squishyplanet` allows for any law that can be cast as a polynomial (e.g. high-order approximations to grid-based models). Second, `ellc` allows for gravity-deformed stars, while `squishyplanet` always models the central star as a sphere and restricts triaxial deformations to the planet only. Third, `ellc` allows users to model radial velocity curves, including the Rossiter-McLaughlin effect, while `squishyplanet` is focused on lightcurve modeling only. In terms of implementation, `ellc` is written in Fortran and wrapped in Python, while `squishyplanet` is written in Python/`JAX`. Also, `ellc` integrates the flux blocked by the planet via 2D numerical integration, while `squishyplanet` uses a 1D numerical integration scheme. We believe that these tools will be complementary and that users will benefit from having both available.
 
 
 # Acknowledgements
