+++ tests/test_oblate_system.py
@@ -16,19 +16,10 @@ def test_limb_darkening_profile_helper():
         poly_order = int(poly_order[0])
         u_coeffs = jax.random.uniform(subkeys[1], (poly_order,), minval=0.0, maxval=1.0)
 
-        state = {
-            "t_peri": 0.0,
-            "times": jnp.linspace(0.0, 5, 10),
-            "a": 2.0,
-            "period": 10,
-            "r": 0.1,
-            "ld_u_coeffs": u_coeffs,
-        }
-
-        system = OblateSystem(**state)
-
         n = 100_000
-        z = system.limb_darkening_profile(jnp.linspace(0, 1, n))
+        z = OblateSystem.limb_darkening_profile(
+            ld_u_coeffs=u_coeffs, r=jnp.linspace(0, 1, n)
+        )
         z *= jnp.linspace(0, 1, n)
 
         # the ui changed for jnp in #20524, don't want to limit versions just for
