### 2024-08-13 e6d72e3
further joss fixes
