+++ docs/index.md
@@ -45,11 +45,12 @@ changelog
 :hidden:
 :caption: Tutorials/Demos
 
-tutorials/illustrations.ipynb
-tutorials/lightcurve_compare.ipynb
+tutorials/fit_oblate_transit.ipynb
 tutorials/create_a_lightcurve.ipynb
 tutorials/create_a_phase_curve.ipynb
-tutorials/fit_oblate_transit.ipynb
+tutorials/illustrations.ipynb
+tutorials/lightcurve_compare.ipynb
+
 
 ```
 
