### 2024-08-13 ben-cassese/squishyplanet#10 @ben-cassese
Hi @rferrerc! I've tried to address these comments, let me know what you think and if there's more to improve on these. Here's a summary of the changes:

> Installation: I use a Linux machine, and following the [installation instructions](https://squishyplanet.readthedocs.io/en/latest/installation.html#mac-linux-users) I tried all three options; options 1 and 3 worked well out-of-the-box, but trying option 2 I got the following error: ERROR: Invalid requirement: 'squishyplanet@git:https://github.com/ben-cassese/squishyplanet' Hint: It looks like a path. File 'squishyplanet@git:https://github.com/ben-cassese/squishyplanet' does not exist. . Not sure if the command line showed on option 2 is also intended to be an out-of-the-box line to be run or if previous steps are implied for the source installation, but flagging this as some clarification would be useful.

Thanks to you (and @catrionamurray!) for catching this- I just copied the wrong commands there, but the instructions should be fixed now. Let me know if you still have trouble.

> Example usage: Most of the core methods are amply used in the documentation examples, except for OblateSystem.loglike(). I believe it would be useful to show an example using this method on real or simulated data, since it would also help satisfy the checklist requirement of "including examples of how to use the software (ideally to solve real-world analysis problems)".

I've added a new tutorial to the online documentation, [Fit a transit of an oblate planet](https://squishyplanet.readthedocs.io/en/latest/tutorials/fit_oblate_transit.html). This walks a user through an end-to-end mock analysis project, and along the way touches on most of the `squishyplanet`'s capabilities. It shows how to a) simulate an accurate transit using a high-order polynomial fit to a limb darkening profile generated from a stellar model grid (I also added a convenience function to `OblateSystem` to help with this), b) how to use `OblateSystem.loglike` to obtain a best-fit solution, then c) how to use `squishyplanet` within a nested sampling framework.

> Functionality documentation: I think it would be helpful to define the conventions used for some of the orbital elements (e.g., is an "edge-on" orbit i=0° or i=90°, in which direction does big Omega increase, etc). Either in the API docs or in the geometry visualization section would be helpful.

I've added another section to the [illustrations](https://squishyplanet.readthedocs.io/en/latest/tutorials/illustrations.html) notebook that goes over our angle conventions.

> State of the field: In the second paragraph of the Statement of Need section, it would be helpful to have a few words on how squishyplanet differs from ellc (the code from Maxted 2016) since, from what I understand in the text, Maxted 2016 can also model non-spherical transits.

I've included a footnote in the paper explicitly comparing the two codes. They do share a similar overarching goal, but each has some capabilities the other lacks, and they implement different algorithms in different languages under the hood.

Let me know if these changes address your concerns, or if you have more! Thanks again for volunteering to review.
