### 2024-08-13 ben-cassese/squishyplanet#11 @ben-cassese
Hi @catrionamurray!

I've attempted to address these comments, thank you again for volunteering to review.

> Reproducibility: Not quite original results, however, the figure in the paper is only briefly described with "a more complete description" in the online documentation. I followed the link to the documentation, however, this specific plot is not generated in the docs. I would like to see the plot + full explanation of each light curve in the docs and/or in the paper.

Apologies for this being unclear: the figure in the paper was created with the [figure.py](https://github.com/ben-cassese/squishyplanet/blob/main/figure.py) file in the same directory as the paper. I've changed the link within the paper to now point directly to this file, and added comments to the file to explain each of the light curves.

> Installation: Similar to https://github.com/ben-cassese/squishyplanet/issues/10 I also have an issue with the second installation method. The 1st and 3rd installation methods work fine for me.
Error:

Just mentioned also in #10, I've now fixed the install command but let me know if you still have issues.

> Example usage: The Quickstart tutorial has no documentation. Since this is the user's first introduction to the code it would be nice to have some context and information about running the code here!

Thanks for the suggestion, I was trying to keep the Quickstart quick but it could benefit from more context. I've added that, and also attempted to direct users to the Geometry Visualizations page and other tutorials for more complete explanations.

> Functionality documentation: I think the API has a good depth of detail, however, in general, descriptions of individual steps in the tutorials lack clarity. Some of this would be remedied by a more detailed Quickstart tutorial. Some examples to better illustrate my point:
Illustrations -> "Now let’s flatten it:" - How do you do this? What axis are you flattening? What parameter(s) are important here?
Illustrations -> "That was all for emitted light, but we can visualize the reflected light too:" - What parameter(s) are you changing to do this?
Create a Phase Curve -> No description of the setup under Reflections.

I've addressed these specific examples and added extra context elsewhere. I've also added a new tutorial, [Fit a transit of an oblate planet](https://squishyplanet.readthedocs.io/en/latest/tutorials/fit_oblate_transit.html), that gives users another example of how to use most of the features.

>  The loglike function is not demonstrated in the tutorials.

This is addressed in the new fitting tutorial.

> State of the field: added small follow-up #10 on oblateness of host stars.

I've added a footnote to the paper addressing the similarities and differences between the two codes. `squishyplanet` does not currently model the gravity darkening or deformation of the host star, though tools for that exist elsewhere. `ellc`, as mentioned, is one, and `starry`'s [OblateMaps](https://starry.readthedocs.io/en/latest/OblateMap/?highlight=oblate) is another. The extension to the `starry` algorithm to account for stellar oblateness is described in [Dholakia, Luger, and Dholakia 2022](https://ui.adsabs.harvard.edu/abs/2022ApJ...925..185D/abstract).

Thank you again!
