### 2024-08-14 ben-cassese/squishyplanet#10 @rferrerc
Hi @ben-cassese , thanks for addressing the comments! The fixes look great, the second installation works as intended now and the new tutorial is excellent! I have checked the corresponding items in the checklist. I'm currently putting together some small comments on the functionality section, but might create a separate issue to keep things focused. I think this issue can be closed now.

Thanks!
