### 2024-08-15 ben-cassese/squishyplanet#13 @rferrerc
JOSS review: Functionality and documentation
Hi @ben-cassese et al!

These are the rest of the comments I had about the submission. Thanks for addressing the first ones from [here](https://github.com/ben-cassese/squishyplanet/issues/10). Most of the submission looks great; from following the tutorials and testing the code everything seems to work properly. These are some minor comments and suggestions.

**Functionality:**

- When running almost anything with `squishyplanet` (e.g. when using the `.lightcurve()` method) I get the following warning from `jax`:

> `/.../python3.12/site-packages/jax/_src/core.py:691: FutureWarning: unhashable type: <class 'jax._src.interpreters.partial_eval.DynamicJaxprTracer'>. Attempting to hash a tracer will lead to an error in a future JAX release. warnings.warn(`

The warning does not produce an error or affect the functioning of the code as far as I can see, but I wonder if it's possible to change what is causing this warning or (probably more easily) to restrict the `jax `version in the requirements to avoid future incompatibility.

**Documentation:**

- In the "Limb darkening profile check" of the _Ilustrations_ notebook/tutorial, the variable `radii` is used but not previously defined; probably just a miss-spell for `stellar_radii`.
- In the _Create a transit lightcurve_, the attribute `projected_effective_r` is wrongly called as `effective_projected_r`. Similarly, in the same notebook at one point the state dictionary is created with a key called `projected_r` which does not exist (should it be `r` or `projected_effective_r` instead?)
- In the _Fit a transit of an oblate planet_ tutorial, add a full reference or ADS link to Kipping 2013 when it is first mentioned.
- At the bottom of the documentation there's a legend that reads "By Author name not set".

**Software paper:**

- State of the field/statement of need: The newly added footnote properly comments on some differences between `ellc` and `squishyplanet`. One good thing to add/change in my opinion is a sentence or two on "advantages" that `squishyplanet` has over `ellc` or other software, or limitations that `squishyplanet` does not have but others do. I see for example that `squishyplanet` "allows for any [limb darkening] law that can be cast as a polynomial". I wonder if there are other advantages or other value that it adds  (including non-functional matters like is it faster/more user friendly/easier to debug/lighter weight? etc.) that could be emphasized when software like e.g. `ellc` already exists. I believe that would be a more informative comparison to the community that, say, what back-end is it written on, when it comes to the community deciding to use the software for their specific cases.

Aside from the comments above, one general small comment would be that the new tutorial on Fitting a transit of an oblate planet (great addition!) could be more accessible for non-specialists by adding just a few words or references/links before certain cells, if anything just to explain what the code cell below is doing. Some examples:

- Output of the 2nd cell: Just mentioning by name what the Greek letter mu represents in the paragraph above.
- Output of 8th cell: Reminding the user that the first plot is the contour of the planet, since it looks different from the illustrations in the Illustrations section.
- Output of the 10th cell: Explicitly mentioning what short and long cadences refer to; saying in the paragraph that 0,1,2,3 refer to polynomial orders (the labels "short 0, short 1" etc. were a bit confusing at first); adding a few words with something to the effect of "...in order to do X, let's look at the Lomb-Scargle periodogram" or something like that, and adding a link to Wikipedia or some paper about Lomb-Scargle.
- Output of the 16th cell: A few words in the paragraph above or comments in the code box saying what changes made are fixing the oblateness to zero in the system and optimization; is there a rationale behind the bounds imposed in the minimization routine for each parameter?
- Output of the 17th cell: Just a sentence being more explicit about how the models are being compared would be useful. Something like "To compare the models, let's use the Bayesian Information Criterion (BIC, link to reference/Wikipedia) and the Akaike Information Criterion (AIC, link to reference/Wikipedia)" or similar.
- In general, adding short comments throughout the code just stating what a piece of code is doing would be useful (like it was done in the 18th cell of this tutorial, the first of the Nested Sampling Section).

Please reach out for any questions or clarifications about the comments above. The paper, software and documentation all look great! Looking forward to finish the review soon.

Thanks!

[Review issue](https://github.com/openjournals/joss-reviews/issues/6972)
