### 2024-08-15 ben-cassese/squishyplanet#13 @ben-cassese
Thanks for the additional comments! Here are my responses to the first two sections:

### Functionality:

> When running almost anything with squishyplanet (e.g. when using the .lightcurve() method) I get the following warning from jax:

Thanks for the catch! This looks like it's caused by a [very recent change](https://github.com/google/jax/blob/82d3cfb3c6f88321f0b29b4cc41134a464de82c2/jax/_src/core.py#L685) to jax. I've limited the jax and jaxlib versions to <0.4.29 to avoid the warning and hopefully avoid an error with future jax releases, but I might need to revisit the restriction later if inference packages in the jax ecosystem eventually outpace these versions. If you are using 0.4.29-0.4.31 and want to suppress the warning, `warnings.simplefilter` can help:
```python
import warnings
with warnings.catch_warnings():
    warnings.simplefilter("ignore")
    lc = planet.lightcurve()
```

### Documentation:

> In the "Limb darkening profile check" of the Ilustrations notebook/tutorial, the variable radii is used but not previously defined; probably just a miss-spell for stellar_radii.

Thanks, fixed! Incorrectly left this after making changes to `OblateSystem.limb_darkening_profile()`

> In the Create a transit lightcurve, the attribute projected_effective_r is wrongly called as effective_projected_r. Similarly, in the same notebook at one point the state dictionary is created with a key called projected_r which does not exist (should it be r or projected_effective_r instead?)

Ack whoops- I recently changed the way we handle 2D projections and just forgot to update this tutorial. Now instead of `projected_r`, which represented the major axis of the projected ellipse, we use `projected_effective_r`, which represents the radius of a circle with the equivalent area as the ellipse. That helps in a few ways: first, when fitting a transit, the old parametrization led to strong degeneracy between `projected_r` and `projected_f`. That's because the combination of both set the total area, which set the transit depth, which had a much stronger influence on the light curve than the slight deviations set by non-sphericity. In this new parameterization the degeneracy is much weaker. Second, we've now standardized the `state` attributes of `OblateSystems` parameterized by 3D or 2D geometries. That means a planet created under 3D now has a `projected_effective_r` and `projected_theta` that are computed during `__init__` that can be directly applied to a 2D planet. I've simplified the cell you call out to reflect this and cut out the `poly_to_parametric_helper` import. The change is also currently commented out in the changelog and will be added when we release a new version to pypi.

> In the Fit a transit of an oblate planet tutorial, add a full reference or ADS link to Kipping 2013 when it is first mentioned.

Added.

> At the bottom of the documentation there's a legend that reads "By Author name not set".

Removed.
