+++ pyproject.toml
@@ -8,8 +8,8 @@ version = "0.1.2"
 requires-python = ">=3.9"
 dependencies = [
     "requests",
-    "jax",
-    "jaxlib",
+    "jax<0.4.29", # avoiding https://github.com/google/jax/issues/21824 for now
+    "jaxlib<0.4.29",
     "quadax",
     "tqdm",
     "black",
