### 2024-08-15 8780222
fit tutorial edits suggested in #13
