### 2024-08-15 ben-cassese/squishyplanet#13 @ben-cassese
And here are some edits to the new tutorial:

> Output of the 2nd cell: Just mentioning by name what the Greek letter mu represents in the paragraph above.

Added.

> Output of 8th cell: Reminding the user that the first plot is the contour of the planet, since it looks different from the illustrations in the Illustrations section.

Added.

> Output of the 10th cell: Explicitly mentioning what short and long cadences refer to; saying in the paragraph that 0,1,2,3 refer to polynomial orders (the labels "short 0, short 1" etc. were a bit confusing at first); adding a few words with something to the effect of "...in order to do X, let's look at the Lomb-Scargle periodogram" or something like that, and adding a link to Wikipedia or some paper about Lomb-Scargle.

Added an extra paragraph to the preceding cell, and changed the figure legend to be more descriptive.

> Output of the 16th cell: A few words in the paragraph above or comments in the code box saying what changes made are fixing the oblateness to zero in the system and optimization; is there a rationale behind the bounds imposed in the minimization routine for each parameter?

Added an explanation for each bound as a comment to the code. They're all just wider than the priors used in the nested sampling section, but included to keep the minimizer from running off to something unphysical.

> Output of the 17th cell: Just a sentence being more explicit about how the models are being compared would be useful. Something like "To compare the models, let's use the Bayesian Information Criterion (BIC, link to reference/Wikipedia) and the Akaike Information Criterion (AIC, link to reference/Wikipedia)" or similar.

Added the links to the preceding paragraph.

> In general, adding short comments throughout the code just stating what a piece of code is doing would be useful (like it was done in the 18th cell of this tutorial, the first of the Nested Sampling Section).

I've sprinkled in more comments in most of the cells, but let me know if there are still areas that could use clarification.

Thanks for such specific edits on this! Appreciate the time/care on this review, the comments are definitely helping make the tutorials clearer 🙂
