### 2024-08-15 ben-cassese/squishyplanet#13 @ben-cassese
Last, about your comment on the state of the field section: respectfully, I'd rather not explicitly claim one package has an advantage over another in the way you suggest, since which to use really depends on your use case. If you are dealing with binary stars, anything to do with radial velocity, or want to model non-spherical stars, ellc is probably the way to go right now. If you want to simulate or fit a complex limb darkening law, or if you want to use any of the inference packages that rely on your model being written in jax (e.g. [numpyro](https://num.pyro.ai/en/stable/index.html), [jaxns](https://jaxns.readthedocs.io/en/latest/index.html), or [blackjax](https://blackjax-devs.github.io/blackjax/)), then squishyplanet would be more appropriate. The main reason for noting the backends is that the language choice can affect the available ecosystem for the rest of your analysis. We think that squishyplanet fills at least two niches that is not currently filled by any other software (extension and implementation of the [Agol, Luger, and Foreman-Mackey 2020](https://ui.adsabs.harvard.edu/abs/2020AJ....159..123A/abstract) to non-spherical planets, and implementation of a non-spherical planet transit simulator in jax), but that other software is more appropriate for other use cases. I hope that makes sense, let me know if you have other thoughts on this.
