### 2024-08-15 ben-cassese/squishyplanet#13 @rferrerc
Hi @ben-cassese ; I've looked through the changes and everything looks good. The bottom of the documentation still reads "By Author name not set" but that's just a small comment and does not affect the submission checklist. I'll finish checking the checklist items now and finish the review. This issue can be closed now. Great software!
