### 2024-08-15 ben-cassese/squishyplanet#13 @ben-cassese
Strange, I'll look more into the author name later, since it's gone on my local build but still being added to the readthedocs. Otherwise, thanks again for your review!
