### 2024-08-19 #6972 @mbobra
👋 @catrionamurray Do you think the submitting author sufficiently addressed ben-cassese/squishyplanet/issues/11?
