### 2024-08-19 ben-cassese/squishyplanet#11 @catrionamurray
Thanks for the updates @ben-cassese, it looks much better now - especially the docs! I can confirm that almost all of the tutorial notebooks work perfectly for me. I only have 1 small issue that came up:

1) Import Error in Fit Oblate Transit Doc
This may be to do with the Jax version I have?
