### 2024-08-19 #6972 @catrionamurray
I have updated the checklist and posted a small error in the same Github Issue: https://github.com/ben-cassese/squishyplanet/issues/11
