### 2024-08-25 ben-cassese/squishyplanet#11 @ben-cassese
Hi @catrionamurray, and sorry for the delay!

Yes, that's likely caused by jax version issues, since the export module was just recently moved out of jax.experimental. Luckily though, this import isn't actually necessary for this tutorial or for any part of squishyplanet, and it's there only because I forgot to remove it after finalizing the tutorial. I've deleted this line now, so hopefully that removes your error!
