### 2024-08-25 #6972 @ben-cassese
Hi all! Just wanted to note that I've now resolved that last issue, and with it I think all of @catrionamurray's major concerns. Let me know if that's correct, or if there's anything more I can address. Thanks!
