### 2024-08-26 #6972 @mbobra
Congratulations, @ben-cassese -- we're almost ready to publish the paper! Could you please cut a release of `squishyplanet` (please use whatever versioning system you wish), [deposit it in Zenodo](https://guides.github.com/activities/citable-code/), and put the link to the Zenodo deposit in this thread? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.
