+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "squishyplanet"
-version = "0.1.2"
+version = "0.2.0"
 requires-python = ">=3.9"
 dependencies = [
     "requests",
