### 2024-08-26 #6972 @ben-cassese
Thank you so much @catrionamurray and @rferrerc for your reviews! We appreciate all the time you spent helping improve this package.

Thanks to you as well @mbobra! This was a great first experience submitting to JOSS. The deposit is here: https://zenodo.org/records/13377037. Let me know if there's anything else I should provide!
