### 2024-08-27 #6972 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
