### 2024-08-27 #6972 @mbobra
Thank you both so very much for your thorough reviews, @rferrerc and @catrionamurray! Your volunteer time keeps JOSS running and I sincerely appreciate it  ☀️

And thank you for your submission and all the edits, @ben-cassese -- I hope you feel it improved the submission. This is a nice paper!
