### 2024-08-30 #6972 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
