### 2024-09-03 #6972 @Kevin-Mattheus-Moerman
@openjournals/dev I noticed the online rendering shows the title with html keywords: e.g. `<code>squishyplanet</code>` see here: https://joss.theoj.org/papers/10.21105/joss.06972. The authors have used the following to define the title, which seems fine:
```
'`squishyplanet`: modeling transits of non-spherical exoplanets in JAX'
```
We should probably fix this on the JOSS website.
